rpc: refactor into TestchainRPC and SidechainRPC
SidechainRPC contains functions that all sidechains implement, such as
getting the sidechain-balance, sending coins, receiving coins etc.

The scope will probably reduce as soon as we
add the actual though.
