+++ lib/rpc/rpc_testchain.dart
@@ -1,0 +1,302 @@
+import 'dart:async';
+
+import 'package:dart_coin_rpc/dart_coin_rpc.dart';
+import 'package:dio/dio.dart';
+import 'package:sidesail/config/sidechains.dart';
+import 'package:sidesail/logger.dart';
+import 'package:sidesail/pages/tabs/settings_tab.dart';
+import 'package:sidesail/rpc/models/bmm_result.dart';
+import 'package:sidesail/rpc/models/core_transaction.dart';
+import 'package:sidesail/rpc/models/raw_transaction.dart';
+import 'package:sidesail/rpc/rpc_config.dart';
+import 'package:sidesail/rpc/rpc_rawtx.dart';
+import 'package:sidesail/rpc/rpc_sidechain.dart';
+import 'package:sidesail/rpc/rpc_withdrawal_bundle.dart';
+
+/// RPC connection the sidechain node.
+abstract class TestchainRPC extends SidechainRPC {
+  Future<BmmResult> refreshBMM(int bidSatoshis);
+  Future<WithdrawalBundle> mainCurrentWithdrawalBundle();
+  Future<FutureWithdrawalBundle> mainNextWithdrawalBundle();
+}
+
+class TestchainRPCLive extends TestchainRPC {
+  RPCClient? _client;
+
+  // responsible for pinging the node every x seconds,
+  // so we can update the UI immediately when the values change
+  Timer? _connectionTimer;
+
+  // hacky way to create an async class
+  // https://stackoverflow.com/a/59304510
+  TestchainRPCLive._create() {
+    chain = TestSidechain();
+  }
+
+  static Future<TestchainRPCLive> create() async {
+    final rpc = TestchainRPCLive._create();
+    await rpc._init();
+    return rpc;
+  }
+
+  Future<void> _init() async {
+    final config = await readRpcConfig(testchainDatadir(), 'testchain.conf');
+    connectionSettings = SingleNodeConnectionSettings(
+      config.path,
+      config.host,
+      config.port,
+      config.username,
+      config.password,
+    );
+    await createClient();
+    await testConnection();
+    _connectionTimer = Timer.periodic(const Duration(seconds: 10), (timer) async {
+      await testConnection();
+    });
+  }
+
+  @override
+  Future<void> createClient() async {
+    _client = RPCClient(
+      host: connectionSettings.host,
+      port: connectionSettings.port,
+      username: connectionSettings.username,
+      password: connectionSettings.password,
+      useSSL: connectionSettings.ssl,
+    );
+
+    // Completely empty client, with no retry logic.
+    _client!.dioClient = Dio();
+
+    log.i('created client');
+  }
+
+  @override
+  Future<(double, double)> getBalance() async {
+    final confirmed = await _client?.call('getbalance') as double;
+    final unconfirmed = await _client?.call('getunconfirmedbalance') as double;
+
+    return (confirmed, unconfirmed);
+  }
+
+  @override
+  Future<BmmResult> refreshBMM(int bidSatoshis) async {
+    final res = await _client?.call('refreshbmm', [bidSatoshis / 100000000]) as Map<String, dynamic>;
+
+    return BmmResult.fromJson(res);
+  }
+
+  @override
+  Future<String> mainSend(
+    String address,
+    double amount,
+    double sidechainFee,
+    double mainchainFee,
+  ) async {
+    // 1. Get refund address for the sidechain withdrawal. This can be any address we control on the SC.
+    final refund = await _getRefundAddress();
+    log.d('got refund address: $refund');
+
+    final withdrawalTxid = await _client?.call('createwithdrawal', [
+      address,
+      refund,
+      amount,
+      sidechainFee,
+      mainchainFee,
+    ]);
+
+    log.d('created peg-out: ${withdrawalTxid['txid']}');
+
+    return withdrawalTxid['txid'];
+  }
+
+  @override
+  Future<String> mainGenerateAddress() async {
+    var address = await _client?.call('getnewaddress', ['Sidechain Peg In', 'legacy']);
+
+    // This is actually just rather simple stuff. Should be able to
+    // do this client side! Just needs the sidechain number, and we're
+    // off to the races.
+    var formatted = await _client?.call('formatdepositaddress', [address as String]);
+
+    return formatted as String;
+  }
+
+  @override
+  Future<String> sideGenerateAddress() async {
+    var address = await _client?.call('getnewaddress', ['Sidechain Deposit']);
+    return address as String;
+  }
+
+  Future<String> _getRefundAddress() async {
+    var address = await _client?.call('getnewaddress', ['Sidechain Deposit', 'legacy']) as String;
+    return address;
+  }
+
+  @override
+  Future<double> sideEstimateFee() async {
+    final estimate = await _client?.call('estimatesmartfee', [6]) as Map<String, dynamic>;
+    if (estimate.containsKey('errors')) {
+      // 10 sats/byte
+      return 0.001;
+    }
+
+    final btcPerKb = estimate['feerate'] as double;
+
+    // who knows!
+    const kbyteInTx = 5;
+    return btcPerKb * kbyteInTx;
+  }
+
+  @override
+  Future<dynamic> callRAW(String method, [dynamic params]) async {
+    return _client?.call(method, params).catchError((err) {
+      log.t('rpc: $method threw exception: $err');
+      throw err;
+    });
+  }
+
+  @override
+  Future<String> sideSend(String address, double amount, bool subtractFeeFromAmount) async {
+    final withdrawalTxid = await _client?.call('sendtoaddress', [
+      address,
+      amount,
+      '',
+      '',
+      subtractFeeFromAmount,
+    ]);
+
+    return withdrawalTxid;
+  }
+
+  @override
+  Future<List<CoreTransaction>> listTransactions() async {
+    // first list
+    final transactionsJSON = await _client?.call('listtransactions', [
+      '',
+      9999, // how many txs to list. We have not implemented pagination, so we list all
+    ]) as List<dynamic>;
+
+    // then convert to something other than json
+    List<CoreTransaction> transactions = transactionsJSON.map((jsonItem) => CoreTransaction.fromMap(jsonItem)).toList();
+    transactions.removeWhere((t) => t.amount == 0);
+    return transactions;
+  }
+
+  @override
+  Future<int> mainBlockCount() async {
+    final cached = await _client?.call('updatemainblockcache') as Map<String, dynamic>;
+
+    return cached['cachesize'];
+  }
+
+  @override
+  Future<int> sideBlockCount() async {
+    return await _client?.call('getblockcount');
+  }
+
+  @override
+  Future<void> ping() async {
+    await _client?.call('ping') as Map<String, dynamic>?;
+  }
+
+  @override
+  void dispose() {
+    _connectionTimer?.cancel();
+    super.dispose();
+  }
+
+  @override
+  Future<WithdrawalBundle> mainCurrentWithdrawalBundle() async {
+    final rawWithdrawalBundle = await _client?.call('getwithdrawalbundle');
+
+    final decoded = await _client?.call('decoderawtransaction', [rawWithdrawalBundle]);
+    final tx = RawTransaction.fromJson(decoded);
+
+    final info = await _client?.call(
+      'getwithdrawalbundleinfo',
+      [tx.hash],
+    );
+
+    final withdrawalIDs = info['withdrawals'] as List<dynamic>;
+
+    final withdrawals = await Future.wait(
+      withdrawalIDs.map(
+        (id) => _client!.call(
+          'getwithdrawal',
+          [id],
+        ).then((json) => Withdrawal.fromJson(json)),
+      ),
+    );
+
+    return WithdrawalBundle.fromRawTransaction(
+      tx,
+      BundleInfo.fromJson(info),
+      withdrawals,
+    );
+  }
+
+  @override
+  Future<FutureWithdrawalBundle> mainNextWithdrawalBundle() async {
+    final rawNextBundle = await _client?.call('listnextbundlewithdrawals') as List<dynamic>;
+
+    return FutureWithdrawalBundle(
+      cumulativeWeight: 0, // TODO: not sure how to obtain this
+      withdrawals: rawNextBundle
+          .map(
+            (withdrawal) => Withdrawal(
+              mainchainFeesSatoshi: withdrawal['amountmainchainfee'],
+              amountSatoshi: withdrawal['amount'],
+              address: withdrawal['destination'],
+              hashBlindTx: '', // TODO
+              refundDestination: '', // TODO
+              status: '', // TODO
+            ),
+          )
+          .toList(),
+    );
+  }
+}
+
+class TestchainRPCError {
+  static const errNoWithdrawalBundle = -100;
+  static const errWithdrawalNotFound = -101;
+}
+
+class RPCError {
+  static const errMisc = -3;
+  static const errNoWithdrawalBundle = -100;
+  static const errWithdrawalNotFound = -101;
+}
+
+class BundleInfo {
+  final int amountSatoshi;
+  final int feesSatoshi;
+  final int weight;
+  final int height;
+
+  BundleInfo({
+    required this.amountSatoshi,
+    required this.feesSatoshi,
+    required this.weight,
+    required this.height,
+  });
+
+  factory BundleInfo.fromJson(Map<String, dynamic> json) {
+    return BundleInfo(
+      amountSatoshi: json['amount'],
+      feesSatoshi: json['fees'],
+      weight: json['weight'],
+      height: json['height'],
+    );
+  }
+
+  Map<String, dynamic> toJson() {
+    return {
+      'amount': amountSatoshi,
+      'fees': feesSatoshi,
+      'weight': weight,
+      'height': height,
+    };
+  }
+}
